pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "streamlit>=1.42.0",
]
